requirements.txt
absl-py==1.4.0
accelerate==0.21.0
addict==2.4.0
aiohttp==3.8.5
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
anyio==3.7.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
augmentation==0.7
Babel==2.12.1
backcall==0.2.0
basicsr==1.4.2
beautifulsoup4==4.12.2
bleach==6.0.0
boltons==23.0.0
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
clean-fid==0.1.35
clip==0.2.0
colorama==0.4.6
comm==0.1.4
contourpy==1.1.0
cycler==0.11.0
debugpy==1.6.7.post1
decorator==5.1.1
defusedxml==0.7.1
docopt==0.6.2
einops==0.6.1
exceptiongroup==1.1.3
executing==1.2.0
fastjsonschema==2.18.0
ffmpeg
ffprobe
filelock==3.12.2
flatbuffers==23.5.26
fonttools==4.42.1
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.6.0
ftfy==6.1.1
future==0.18.3
google-auth==2.22.0
google-auth-oauthlib==1.0.0
grpcio==1.57.0
huggingface-hub==0.16.4
idna==3.4
imageio==2.31.1
importlib-metadata==6.8.0
ipykernel==6.25.1
ipython==8.14.0
ipython-genutils==0.2.0
ipywidgets==7.7.1
isoduration==20.11.0
jedi==0.19.0
Jinja2==3.1.2
json5==0.9.14
jsonmerge==1.9.2
jsonpointer==2.4
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter_server==2.7.2
jupyter_server_terminals==0.4.4
jupyterlab==4.0.5
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.8
jupyterlab_server==2.24.0
kiwisolver==1.4.4
kornia==0.7.0
lark==1.1.7
lazy_loader==0.3
lightning-utilities==0.9.0
lmdb==1.4.1
lpips==0.1.4
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mediapipe==0.10.3
mistune==3.0.1
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
nbclient==0.8.0
nbconvert==7.7.4
nbformat==5.9.2
nest-asyncio==1.5.7
networkx==3.1
notebook==7.0.2
notebook_shim==0.2.3
numpy==1.25.2
oauthlib==3.2.2
omegaconf==2.3.0
open-clip-torch==2.20.0
opencv-contrib-python==4.8.0.76
opencv-python==4.8.0.76
overrides==7.4.0
packaging==23.1
pandas==2.0.3
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
piexif==1.1.3
Pillow==10.0.0
platformdirs==3.10.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==3.20.3
psutil==5.9.5
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.16.1
pyparsing==3.0.9
pyre-extensions==0.0.29
python-dateutil==2.8.2
python-json-logger==2.0.7
pytorch-lightning==1.8.4
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0.1
pyzmq==25.1.1
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.9.2
rsa==4.9
runpod==1.2.0
safetensors==0.3.2
scikit-image==0.21.0
scipy==1.11.2
Send2Trash==1.8.2
sentencepiece==0.1.99
six==1.16.0
sniffio==1.3.0
sounddevice==0.4.6
soupsieve==2.4.1
stack-data==0.6.2
sympy==1.12
tb-nightly==2.15.0a20230822
tensorboard-data-server==0.7.1
tensorboardX==2.6.2.2
terminado==0.17.1
tifffile==2023.8.12
timm==0.9.5
tinycss2==1.2.1
tokenizers==0.13.3
tomli==2.0.1
torchdiffeq==0.2.3
torchmetrics==0.11.4
torchsde==0.2.5
tornado==6.3.3
tqdm==4.66.1
traitlets==5.9.0
trampoline==0.1.2
transformers==4.32.0
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
uri-template==1.3.0
urllib3
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.2
Werkzeug==2.3.7
wget==3.2
widgetsnbextension==3.6.5
xformers==0.0.20
yapf==0.40.1
yarl==1.9.2
zipp==3.16.2

libjpeg-dev
libpng-dev